use a local cache to avoid download the binaries for every build

+++ src/main/java/io/jenkins/plugins/adoptopenjdk/AdoptOpenJDKInstaller.java
@@ -39,13 +39,21 @@ import hudson.tools.ToolInstallation;
 import hudson.tools.ToolInstaller;
 import hudson.tools.ToolInstallerDescriptor;
 import hudson.tools.ZipExtractionInstaller;
+import jenkins.model.Jenkins;
 import jenkins.security.MasterToSlaveCallable;
 import net.sf.json.JSONObject;
+import org.apache.commons.io.IOUtils;
 import org.jenkinsci.Symbol;
 import org.kohsuke.stapler.DataBoundConstructor;
 
 import javax.annotation.Nonnull;
+import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
@@ -55,6 +63,8 @@ import java.util.Locale;
  * Based on https://github.com/jenkinsci/jdk-tool-plugin
  */
 public class AdoptOpenJDKInstaller extends ToolInstaller {
+
+    private static boolean DISABLE_CACHE = Boolean.getBoolean( AdoptOpenJDKInstaller.class + ".cache.disable" );
 
     /**
      * AdoptOpenJDK release id
@@ -105,24 +115,50 @@ public class AdoptOpenJDKInstaller extends ToolInstaller {
             if (binary == null) {
                 throw new IOException(Messages.AdoptOpenJDKInstaller_performInstallation_binaryNotFound(id, p.name(), c.name()));
             }
-            String url = binary.binary_link;
-
-            ZipExtractionInstaller zipExtractionInstaller = new ZipExtractionInstaller(null, url, null);
-            FilePath installation = zipExtractionInstaller.performInstallation(tool, node, log);
-
-            installation.child(".timestamp").delete(); // we don't use the timestamp
-            FilePath base = findPullUpDirectory(installation, p);
-            if (base != null && base != expected) {
-                base.moveAllChildrenTo(expected);
+            File cache = getLocalCacheFile(p, c);
+            if (!DISABLE_CACHE && cache.exists() && cache.length()>1*1024*1024) { // if the file is too small, don't trust it
+                // the zip contains already the directory so we unzip to parent directory
+                expected.getParent().installIfNecessaryFrom(cache.toURI().toURL(), log,
+                                                            Messages.AdoptOpenJDKInstaller_performInstallation_path(expected));
+                expected.getParent().child( ".timestamp" ).delete();
+            } else {
+                String url = binary.binary_link;
+                ZipExtractionInstaller zipExtractionInstaller = new ZipExtractionInstaller(null, url, null);
+                FilePath installation = zipExtractionInstaller.performInstallation(tool, node, log);
+                installation.child(".timestamp").delete(); // we don't use the timestamp
+                FilePath base = findPullUpDirectory(installation, p);
+                if (base != null && base != expected) {
+                    base.moveAllChildrenTo(expected);
+                }
+                marker.write(id, null);
+                if(!DISABLE_CACHE) {
+                    // update the local cache on master
+                    // download to a temporary file and rename it in to handle concurrency and failure correctly,
+                    Path tmp = new File( cache.getPath()+".tmp").toPath();
+                    try {
+                        Path tmpParent = tmp.getParent();
+                        if (tmpParent != null) {
+                            Files.createDirectories(tmpParent);
+                        }
+                        try (OutputStream out = Files.newOutputStream(tmp)) {
+                            expected.zip(out);
+                        }
+                        Files.move(tmp, cache.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                    } finally {
+                        Files.deleteIfExists(tmp);
+                    }
+                }
             }
-            marker.write(id, null);
-
-            return installation;
         } catch (DetectionFailedException e) {
             log.getLogger().println(Messages.AdoptOpenJDKInstaller_performInstallation_JdkSkipped(e.getMessage()));
         }
 
         return expected;
+    }
+
+    private File getLocalCacheFile( Platform platform, CPU cpu) {
+        // we force .zip file
+        return new File(Jenkins.get().getRootDir(), "caches/adoptopenjdk/"+platform+"/"+cpu+"/"+id+".zip");
     }
 
     /**
